+++ config/metric_test.go
@@ -4,15 +4,19 @@
 package config
 
 import (
+	"context"
+	"errors"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/noop"
+	"go.opentelemetry.io/otel/sdk/resource"
 )
 
-func TestInitMeterProvider(t *testing.T) {
+func TestMeterProvider(t *testing.T) {
 	tests := []struct {
 		name         string
 		cfg          configOptions
@@ -23,10 +27,53 @@ func TestInitMeterProvider(t *testing.T) {
 			name:         "no-meter-provider-configured",
 			wantProvider: noop.NewMeterProvider(),
 		},
+		{
+			name: "error-in-config",
+			cfg: configOptions{
+				opentelemetryConfig: OpenTelemetryConfiguration{
+					MeterProvider: &MeterProvider{
+						Readers: []MetricReader{
+							{
+								Periodic: &PeriodicMetricReader{},
+								Pull:     &PullMetricReader{},
+							},
+						},
+					},
+				},
+			},
+			wantProvider: noop.NewMeterProvider(),
+			wantErr:      errors.Join(errors.New("must not specify multiple metric reader type")),
+		},
+		{
+			name: "multiple-errors-in-config",
+			cfg: configOptions{
+				opentelemetryConfig: OpenTelemetryConfiguration{
+					MeterProvider: &MeterProvider{
+						Readers: []MetricReader{
+							{
+								Periodic: &PeriodicMetricReader{},
+								Pull:     &PullMetricReader{},
+							},
+							{
+								Periodic: &PeriodicMetricReader{
+									Exporter: MetricExporter{
+										Console: Console{},
+										OTLP:    &OTLPMetric{},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+			wantProvider: noop.NewMeterProvider(),
+			wantErr:      errors.Join(errors.New("must not specify multiple metric reader type"), errors.New("must not specify multiple exporters")),
+		},
 	}
 	for _, tt := range tests {
-		mp, err := initMeterProvider(tt.cfg)
+		mp, shutdown, err := meterProvider(tt.cfg, resource.Default())
 		require.Equal(t, tt.wantProvider, mp)
-		require.NoError(t, tt.wantErr, err)
+		assert.Equal(t, tt.wantErr, err)
+		require.NoError(t, shutdown(context.Background()))
 	}
 }